Rename codec.rs to message_codec.rs
